# api-docs/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
...
  /api/plans/edited:
    put:
      tags:
      - rest-plan
      operationId: editPlan
      parameters:
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: newDescription
        in: query
        required: true
        schema:
          type: string
      - name: abv
        in: query
        required: true
        schema:
          type: string
      - name: difficulty
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not Found
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/categories/edit:
    put:
      tags:
      - rest-category
      summary: Edits an existing category
      operationId: editCategory
      parameters:
      - name: categoryName
        in: query
        required: true
        schema:
          type: string
      - name: newDescription
        in: query
        required: true
        schema:
          type: string
      - name: color
        in: query
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                imageFile:
                  type: string
                  format: binary
      responses:
        "404":
          description: Not Found
        "200":
          description: Returns the created category information
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
        "403":
          description: Only access to admin account
        "409":
          description: There is already a category with the name given by parameter
...
  /api/plans/new:
    post:
      tags:
      - rest-plan
      operationId: createPlan
      parameters:
      - name: category
        in: query
        required: true
        schema:
          type: string
      - name: planName
        in: query
        required: true
        schema:
          type: string
      - name: abv
        in: query
        required: true
        schema:
          type: string
      - name: description
        in: query
        required: true
        schema:
          type: string
      - name: difficulty
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "404":
          description: Not Found
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/Plan_PlanDetails'
...
  /api/users/profile:
    get:
      tags:
      - rest-user
      summary: Get user profile by the email
      operationId: getUser
      responses:
        "404":
          description: Not Found
        "200":
          description: Found the user profile
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/User_UserDetails'
        "403":
          description: Only access to registered user
...
  /api/categories:
    get:
      tags:
      - rest-category
      summary: Get the information a specific category searched by its name
      operationId: categoryInfo
      parameters:
      - name: name
        in: query
        required: true
        schema:
          type: string
      responses:
        "404":
          description: Doesn't exists a category with that name
          content: {}
        "200":
          description: Returns the category information
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/CategoryDetails_CategoryDetails'
...
components:
  schemas:
    Category_UserDetails:
      type: object
      properties:
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_UserDetails'
        description:
          type: string
    Tree_UserDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_UserDetails'
        category:
          $ref: '#/components/schemas/Category_UserDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_UserDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_UserDetails'
        email:
          type: string
        username:
          type: string
        name:
          type: string
        surname:
          type: string
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        roles:
          type: array
          items:
            type: string
        password:
          type: string
          writeOnly: true
    Category_PlanDetails:
      type: object
      properties:
        name:
          type: string
        des:
          type: string
        color:
          type: string
        likedByUser:
          type: boolean
        plans:
          type: array
          items:
            $ref: '#/components/schemas/Plan_PlanDetails'
        userFavoritesCategory:
          type: array
          items:
            $ref: '#/components/schemas/User_PlanDetails'
        description:
          type: string
    Plan_PlanDetails:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        difficulty:
          type: integer
          format: int32
        likedUser:
          type: boolean
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
    Tree_PlanDetails:
      type: object
      properties:
        user:
          $ref: '#/components/schemas/User_PlanDetails'
        category:
          $ref: '#/components/schemas/Category_PlanDetails'
        last_update:
          type: integer
          format: int64
        date:
          type: string
        imagePath:
          type: string
    User_PlanDetails:
      type: object
      properties:
        trees:
          type: array
          items:
            $ref: '#/components/schemas/Tree_PlanDetails'
        encodedPassword:
          type: string
        confirmEncodedPassword:
          type: string
        password:
          type: string
          writeOnly: true
...
    CategoryDetails_CategoryDetails:
      type: object
...